Fix [#12] 홈피드불러오기 오류 수정

+++ back/dipping/src/main/java/com/common/dipping/api/board/domain/dto/BoardResponse.java
@@ -15,7 +15,7 @@ public class BoardResponse {
 	private Boolean openPost;
 	private Boolean openComment;
 	private Boolean albumArt;
-	private Long likeCount;
+	private int likeCount;
 	private Boolean myLike;
 	private int commentCount;
